Move Remaining Web Worker Files
